+++ changelog.md
@@ -1,6 +1,8 @@
 ## 2.3.0
 - [FIX][IOS] Fix Auto resize for UIkit table cells
 - [FEA][AND] Improve recipe card view
+- [FEA][AND] Improve recipe detail view recomposition 
+- [FEA][COR] Simplifying Price ViewModel
 
 ## 2.2.6
 - [FIX][AND] Difficulty checkbox in filter are now working